+++ product-microservice/src/main/java/product/microservice/productmicroservice/repository/RatingRepository.java
@@ -3,5 +3,5 @@ package product.microservice.productmicroservice.repository;
 import org.springframework.data.repository.CrudRepository;
 import product.microservice.productmicroservice.model.Rating;
 
-public interface RatingRepository extends CrudRepository<Rating, Integer> {
+public interface RatingRepository extends CrudRepository<Rating, Long> {
 }
